+++ tests/unit/test_utils.py
@@ -1,5 +1,5 @@
 import json
-import pytest
+import logging
 
 import pandas as pd
 
@@ -160,11 +160,12 @@ def test_validate_column_size_pass():
         assert False, "Validation failed but was expected to pass"
 
 
-def test_validate_column_size_fail():
+def test_validate_column_size_fail(caplog):
     df = pd.DataFrame({"col1": ["a", "bb", "cccc"]})
     tests = {"column_size": {"col1": 3}}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "field length is different than 3" in caplog.text
 
 
 def test_validate_column_unique_values_pass():
@@ -176,11 +177,12 @@ def test_validate_column_unique_values_pass():
         assert False, "Validation failed but was expected to pass"
 
 
-def test_validate_column_unique_values_fail():
+def test_validate_column_unique_values_fail(caplog):
     df = pd.DataFrame({"col1": [1, 2, 2]})
     tests = {"column_unique_values": ["col1"]}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "Values for col1 are not unique." in caplog.text
 
 
 def test_validate_column_list_to_match_pass():
@@ -192,11 +194,12 @@ def test_validate_column_list_to_match_pass():
         assert False, "Validation failed but was expected to pass"
 
 
-def test_validate_column_list_to_match_fail():
+def test_validate_column_list_to_match_fail(caplog):
     df = pd.DataFrame({"col1": [1]})
     tests = {"column_list_to_match": ["col1", "col2"]}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "Columns are different than expected" in caplog.text
 
 
 def test_validate_dataset_row_count_pass():
@@ -208,11 +211,12 @@ def test_validate_dataset_row_count_pass():
         assert False, "Validation failed but was expected to pass"
 
 
-def test_validate_dataset_row_count_fail():
+def test_validate_dataset_row_count_fail(caplog):
     df = pd.DataFrame({"col1": [1, 2, 3, 4, 5, 6]})
     tests = {"dataset_row_count": {"min": 1, "max": 5}}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "Row count (6) is not between 1 and 5" in caplog.text
 
 
 def test_validate_column_match_regex_pass():
@@ -224,11 +228,12 @@ def test_validate_column_match_regex_pass():
         assert "Validation failed but was expected to pass"
 
 
-def test_validate_column_match_regex_fail():
+def test_validate_column_match_regex_fail(caplog):
     df = pd.DataFrame({"col1": ["A123", "B34", "C45"]})
     tests = {"column_match_regex": {"col1": "^[A-Z][0-9]{2}$"}}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "[column_match_regex] on col1 column failed!" in caplog.text
 
 
 def test_validate_column_sum_pass():
@@ -240,8 +245,9 @@ def test_validate_column_sum_pass():
         assert False, "Validation failed but was expected to pass"
 
 
-def test_validate_column_sum_fail():
+def test_validate_column_sum_fail(caplog):
     df = pd.DataFrame({"col1": [1, 2, 3, 4]})
     tests = {"column_sum": {"col1": {"min": 5, "max": 6}}}
-    with pytest.raises(ValidationError):
+    with caplog.at_level(logging.INFO):
         validate(df, tests)
+    assert "Sum of 10 for col1 is out of the expected range - <5:6>" in caplog.text